backend/app.py
from flask import Flask, request, jsonify
from azure.cognitiveservices.vision.computervision import ComputerVisionClient
from msrest.authentication import CognitiveServicesCredentials

app = Flask(__name__)

subscription_key = 'your_subscription_key'
endpoint = 'your_endpoint'
computervision_client = ComputerVisionClient(endpoint, CognitiveServicesCredentials(subscription_key))

@app.route('/classify-image', methods=['POST'])
def classify_image():
    file = request.files['file']
    image_data = file.read()
    result = computervision_client.analyze_image_in_stream(image_data)
    return jsonify(result.as_dict())

@app.route('/rag-chatbot', methods=['POST'])
def rag_chatbot():
    query = request.json.get('query')
    # Implement RAG logic here
    response = {"answer": "This is a placeholder response."}
    return jsonify(response)

if __name__ == '__main__':
    app.run(debug=True)
