requirements.txt
flask
azure-cognitiveservices-vision-computervision
msrest
azure-ai-formrecognizer
